Split index maintenance commands
